@@ -18,16 +18,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.1.3",
+    "@angular/animations": "^13.2.0",
     "@angular/cdk": "^13.1.2",
     "@angular/common": "^13.1.3",
     "@angular/compiler": "^13.1.2",
     "@angular/core": "^13.1.2",
-    "@angular/forms": "^13.1.3",
-    "@angular/material": "^13.1.3",
+    "@angular/forms": "^13.2.0",
+    "@angular/material": "^13.2.0",
     "@angular/platform-browser": "^13.1.3",
-    "@angular/platform-browser-dynamic": "^13.1.3",
-    "@angular/router": "^13.1.3",
+    "@angular/platform-browser-dynamic": "^13.2.0",
+    "@angular/router": "^13.2.0",
     "@splitsoftware/splitio-browserjs": "^0.4.0",
     "angular-gridster2": "^13.1.1",
     "ng-click-outside": "^9.0.1",
@@ -39,15 +39,15 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "^0.1301.4",
-    "@angular-devkit/build-angular": "^13.1.4",
+    "@angular-devkit/architect": "^0.1302.0",
+    "@angular-devkit/build-angular": "^13.2.0",
     "@angular-eslint/builder": "^13.0.1",
     "@angular-eslint/eslint-plugin": "^13.0.1",
     "@angular-eslint/eslint-plugin-template": "^13.0.1",
     "@angular-eslint/schematics": "^13.0.1",
     "@angular-eslint/template-parser": "^13.0.1",
-    "@angular/cli": "^13.1.4",
-    "@angular/compiler-cli": "^13.1.3",
+    "@angular/cli": "^13.2.0",
+    "@angular/compiler-cli": "^13.2.0",
     "@compodoc/compodoc": "^1.1.18",
     "@strut-software/eslint-config": "^1.8.0",
     "@strut-software/stylelint-config": "^1.1.1",
@@ -58,7 +58,7 @@
     "@typescript-eslint/parser": "5.10.1",
     "autoprefixer": "^10.4.2",
     "cross-env": "^7.0.3",
-    "eslint": "8.7.0",
+    "eslint": "8.8.0",
     "husky": "^7.0.4",
     "jasmine-core": "~4.0.0",
     "jasmine-spec-reporter": "~7.0.0",
@@ -76,7 +76,7 @@
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^22.0.0",
     "stylelint-scss": "^3.21.0",
-    "tailwindcss": "^3.0.16",
+    "tailwindcss": "^3.0.18",
     "ts-node": "~10.4.0",
     "typescript": "~4.4.4"
   }
