@@ -1,10 +1,11 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { HistoryDrawerComponent } from './history-drawer.component';
-
+import { LoadingIndicatorModule } from 'src/app/shared/loading-indicator/loading-indicator.module';
+import { UserAvatarModule } from 'src/app/shared/user-avatar/user-avatar.module';
 @NgModule({
   declarations: [HistoryDrawerComponent],
-  imports: [CommonModule],
+  imports: [CommonModule, UserAvatarModule, LoadingIndicatorModule],
   exports: [HistoryDrawerComponent],
 })
 export class HistoryDrawerModule {}
