@@ -1324,7 +1324,7 @@ export class MockDocumentService {
     } else {
       const eventDocument: DocumentEvent[] = [
         {
-          date: '2022-01-18T22:13:05.871Z',
+          eventTimeUTC: '2022-01-18T22:13:05.871Z',
           description: 'Event Document #1',
           user: {
             rithmId: '123',
